requirements.txt
git
PyGithub